+++ CHANGELOG.md
@@ -1,5 +1,32 @@
 # CHANGELOG
 
+## v13.1.0 - 2020-06-23
+
+### Libraries affected
+
+`storage_typesafe`
+
+Adds a builder for PrefixTransfer
+
+With config:
+
+```hocon
+source.cloudProvider = "aws"
+destination {
+  cloudProvider = "azure"
+  azure.blobStore.connectionString = "UseDevelopmentStorage=true;"
+}
+```
+
+In code:
+
+```scala
+import uk.ac.wellcome.storage.typesafe.PrefixTransferBuilder
+
+val prefixTransfer = PrefixTransferBuilder.build(config)
+assert(prefixTransfer.isInstanceOf[S3toAzurePrefixTransfer])
+```
+
 ## v13.0.0 - 2020-06-23
 
 ### Libraries affected